Refactor: modify 패키지 구조 변경 - 팀 규칙 반영
